v0.11.119.01+06

+++ css/src/css.ml
@@ -212,7 +212,7 @@ let overflow_y = make_overflow "overflow-y"
 
 let z_index i = create_raw ~field:"z-index" ~value:(Int.to_string i)
 
-let opacity i = create_raw ~field:"opacity" ~value:(Int.to_string i)
+let opacity i = create_raw ~field:"opacity" ~value:(Float.to_string i)
 
 let create_length_field field l =
   create_raw ~field ~value:(Auto_or_length.to_string_css l)
